Moved like logic from LikeMovieViewModel into MovieDetailsViewModel
