--- app/src/main/java/com/klewerro/moviedbapp/presentation/LikeMovieState.kt
@@ -1,13 +1,0 @@
-package com.klewerro.moviedbapp.presentation
-
-data class LikeMovieState(
-    val isMovieLiked: Boolean? = null,
-    val likeChanged: LikeChanged = LikeChanged.Unchanged
-)
-
-sealed class LikeChanged {
-    data object Unchanged : LikeChanged()
-    data class Liked(val movieTitle: String) : LikeChanged()
-    data class LikeRemoved(val movieTitle: String) : LikeChanged()
-    data class Error(val movieTitle: String) : LikeChanged()
-}
