+++ movieDetails/src/main/java/com/klewerro/moviedbapp/movieDetails/presentation/MovieDetailsEvent.kt
@@ -1,0 +1,6 @@
+package com.klewerro.moviedbapp.movieDetails.presentation
+
+sealed class MovieDetailsEvent {
+    data object MovieDetails : MovieDetailsEvent()
+    data object DismissChangedDetails : MovieDetailsEvent()
+}
