+++ movieDetails/src/main/java/com/klewerro/moviedbapp/movieDetails/presentation/MovieDetailsScreen.kt
@@ -2,6 +2,7 @@ package com.klewerro.moviedbapp.movieDetails.presentation
 
 import android.content.res.Configuration
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -14,10 +15,13 @@ import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Star
 import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -26,6 +30,7 @@ import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -35,6 +40,8 @@ import coil.compose.AsyncImage
 import com.klewerro.moviedbapp.core.R
 import com.klewerro.moviedbapp.core.presentation.MovieAppBar
 import com.klewerro.moviedbapp.core.ui.theme.MovieDbAppTheme
+import com.klewerro.moviedbapp.core.ui.theme.gold
+import com.klewerro.moviedbapp.core.ui.theme.goldDark
 import com.klewerro.moviedbapp.core.util.testData.MovieTestData
 import com.klewerro.moviedbapp.movieDetails.presentation.composable.MovieDetailsSection
 import kotlin.math.min
@@ -47,11 +54,50 @@ fun MovieDetailsScreen(
     onBackClick: () -> Unit,
     movieDetailsViewModel: MovieDetailsViewModel = hiltViewModel()
 ) {
+    val context = LocalContext.current
     val state by movieDetailsViewModel.state.collectAsStateWithLifecycle()
+
+    LaunchedEffect(state.likeChanged) {
+        when (val changedObject = state.likeChanged) {
+            LikeChanged.Unchanged -> Unit
+            is LikeChanged.Liked -> {
+                snackbarHostState.showSnackbar(
+                    context.getString(
+                        RCore.string.liked_movie_message,
+                        changedObject.movieTitle
+                    )
+                )
+                movieDetailsViewModel.onEvent(MovieDetailsEvent.DismissChangedDetails)
+            }
+
+            is LikeChanged.LikeRemoved -> {
+                snackbarHostState.showSnackbar(
+                    context.getString(
+                        RCore.string.disliked_movie_message,
+                        changedObject.movieTitle
+                    )
+                )
+                movieDetailsViewModel.onEvent(MovieDetailsEvent.DismissChangedDetails)
+            }
+
+            is LikeChanged.Error -> {
+                snackbarHostState.showSnackbar(
+                    message = context.getString(
+                        RCore.string.error_movie_liking_message,
+                        changedObject.movieTitle
+                    ),
+                    duration = SnackbarDuration.Long
+                )
+                movieDetailsViewModel.onEvent(MovieDetailsEvent.DismissChangedDetails)
+            }
+        }
+    }
+
     MovieDetailsScreenContent(
         movieDetailsState = state,
         snackbarHostState = snackbarHostState,
         onBackClick = onBackClick,
+        onStarClick = { movieDetailsViewModel.onEvent(MovieDetailsEvent.MovieDetails) },
         modifier = modifier.fillMaxSize()
     )
 }
@@ -61,6 +107,7 @@ private fun MovieDetailsScreenContent(
     movieDetailsState: MovieDetailsState,
     snackbarHostState: SnackbarHostState,
     onBackClick: () -> Unit,
+    onStarClick: () -> Unit,
     modifier: Modifier = Modifier
 ) {
     val movie = movieDetailsState.movie
@@ -89,15 +136,15 @@ private fun MovieDetailsScreenContent(
                         contentDescription = stringResource(
                             RCore.string.filled_star_content_description
                         ),
-//                        tint = if (isLiked) {
-//                            if (isSystemInDarkTheme()) gold else goldDark
-//                        } else {
-//                            MaterialTheme.colorScheme.onPrimary
-//                        },
+                        tint = if (movieDetailsState.movie.isLiked) {
+                            if (isSystemInDarkTheme()) gold else goldDark
+                        } else {
+                            MaterialTheme.colorScheme.onPrimary
+                        },
                         modifier = Modifier
                             .size(32.dp)
                             .clickable {
-                                // onLikeIconClick()
+                                onStarClick()
                             }
                     )
                 }
@@ -149,6 +196,7 @@ private fun MovieDetailsScreenPreview() {
             movieDetailsState = state,
             snackbarHostState = remember { SnackbarHostState() },
             onBackClick = {},
+            onStarClick = {},
             modifier = Modifier.fillMaxSize()
         )
     }
